feature/android-mvp:update backend and andriod link and update dto and debug for feedlist and ui page

+++ ToutiaoAndroid/app/src/main/java/com/xuziyi/toutiaoandroid/data/remote/dto/FeedItemDto.kt
@@ -59,5 +59,5 @@ data class FeedItemDto(
 
     // 推荐排序权重
     @SerializedName("weight")
-    val weight: Int = 0
+    val weight: Float = 0f
 )
